adjust account book related interfaces and constants
