+++ src/constants/account_book.ts
@@ -1,27 +1,20 @@
 import { SortBy, SortOrder } from '@/constants/sort';
-import { WORK_TAG } from '@/interfaces/account_book';
+import {
+  WORK_TAG,
+  AGENT_FILING_ROLE,
+  DECLARANT_FILING_METHOD,
+  FILING_FREQUENCY,
+  FILING_METHOD,
+} from '@/interfaces/account_book';
 import { ITeam } from '@/interfaces/team';
 
 export const DEFAULT_SORT_OPTIONS = [{ sortBy: SortBy.CREATED_AT, sortOrder: SortOrder.DESC }];
-
-// Info: (20250418 - Liz) 申報頻率
-export enum FILING_FREQUENCY {
-  BIMONTHLY_FILING = 'BIMONTHLY_FILING',
-  MONTHLY_FILING = 'MONTHLY_FILING',
-}
 
 // Info: (20250418 - Liz) 申報頻率選項的補充說明
 export const FILING_FREQUENCY_OPTIONS = [
   { label: FILING_FREQUENCY.BIMONTHLY_FILING, subLabel: 'EVERY_TWO_MONTHS' },
   { label: FILING_FREQUENCY.MONTHLY_FILING, subLabel: 'EVERY_MONTH' },
 ];
-
-// Info: (20250418 - Liz) 申報方式(總繳種類)
-export enum FILING_METHOD {
-  SINGLE_ENTITY_FILING = 'SINGLE_ENTITY_FILING',
-  CONSOLIDATED_FILING = 'CONSOLIDATED_FILING',
-  INDIVIDUAL_FILING = 'INDIVIDUAL_FILING',
-}
 
 // Info: (20250507 - Liz) 申報方式選項的補充說明
 export const FILING_METHOD_OPTIONS = [
@@ -32,19 +25,6 @@ export const FILING_METHOD_OPTIONS = [
   { label: FILING_METHOD.CONSOLIDATED_FILING, subLabel: 'HEAD_OFFICE_FILES_FOR_ALL_BRANCHES' },
   { label: FILING_METHOD.INDIVIDUAL_FILING, subLabel: 'EACH_UNIT_FILES_SEPARATELY' },
 ];
-
-// Info: (20250418 - Liz) (申報人)申報方式
-export enum DECLARANT_FILING_METHOD {
-  SELF_FILING = 'SELF_FILING',
-  AGENT_FILING = 'AGENT_FILING',
-}
-
-// Info: (20250507 - Liz) 申報代理人的角色有三種：會計師(稅務代理人)、記帳士、記帳及報稅代理人
-export enum AGENT_FILING_ROLE {
-  ACCOUNTANT = 'ACCOUNTANT',
-  BOOKKEEPER = 'BOOKKEEPER',
-  BOOKKEEPER_AND_FILING_AGENT = 'BOOKKEEPER_AND_FILING_AGENT',
-}
 
 // Info: (20250421 - Liz) 第一步驟表單狀態
 export interface Step1FormState {
