+++ src/interfaces/account_book.ts
@@ -1,4 +1,3 @@
-// import { DECLARANT_FILING_METHOD, FILING_FREQUENCY, FILING_METHOD, AGENT_FILING_ROLE } from '@/constants/account_book'; // ToDo: (20250507 - Liz) RC2 新增表單欄位
 import { LocaleKey } from '@/constants/normal_setting';
 import { ITeam } from '@/interfaces/team';
 
@@ -23,6 +22,32 @@ export enum ACCOUNT_BOOK_UPDATE_ACTION {
   UPDATE_TAG = 'updateTag',
 }
 
+// Info: (20250418 - Liz) 申報頻率
+export enum FILING_FREQUENCY {
+  BIMONTHLY_FILING = 'BIMONTHLY_FILING',
+  MONTHLY_FILING = 'MONTHLY_FILING',
+}
+
+// Info: (20250418 - Liz) 申報方式(總繳種類)
+export enum FILING_METHOD {
+  SINGLE_ENTITY_FILING = 'SINGLE_ENTITY_FILING',
+  CONSOLIDATED_FILING = 'CONSOLIDATED_FILING',
+  INDIVIDUAL_FILING = 'INDIVIDUAL_FILING',
+}
+
+// Info: (20250418 - Liz) (申報人)申報方式
+export enum DECLARANT_FILING_METHOD {
+  SELF_FILING = 'SELF_FILING',
+  AGENT_FILING = 'AGENT_FILING',
+}
+
+// Info: (20250507 - Liz) 申報代理人的角色有三種：會計師(稅務代理人)、記帳士、記帳及報稅代理人
+export enum AGENT_FILING_ROLE {
+  ACCOUNTANT = 'ACCOUNTANT',
+  BOOKKEEPER = 'BOOKKEEPER',
+  BOOKKEEPER_AND_FILING_AGENT = 'BOOKKEEPER_AND_FILING_AGENT',
+}
+
 // Info: (20250226 - Liz) 原為 ICompany (因為公司已經改名成帳本)
 export interface IAccountBook {
   id: number;
@@ -38,24 +63,25 @@ export interface IAccountBook {
   isPrivate?: boolean; // Deprecated: (20250423 - Liz) 已棄用
 
   // Info: (20250423 - Liz) RC2 新增表單欄位
-  // responsiblePerson: string; // Info: (20250423 - Liz) 負責人姓名
-  // taxSerialNumber: string; // Info: (20250423 - Liz) 稅籍編號
-  // contactPerson: string; // Info: (20250423 - Liz) 聯絡人姓名
-  // phoneNumber: string; // Info: (20250423 - Liz) 電話號碼
-  // city: string; // Info: (20250423 - Liz) 縣市
-  // district: string; // Info: (20250423 - Liz) 行政區
-  // enteredAddress: string; // Info: (20250423 - Liz) 使用者輸入的地址
+  responsiblePerson: string; // Info: (20250423 - Liz) 負責人姓名
+  taxSerialNumber: string; // Info: (20250423 - Liz) 稅籍編號
+  contactPerson: string; // Info: (20250423 - Liz) 聯絡人姓名
+  phoneNumber: string; // Info: (20250423 - Liz) 電話號碼
+  city: string; // Info: (20250423 - Liz) 縣市
+  district: string; // Info: (20250423 - Liz) 行政區
+  enteredAddress: string; // Info: (20250423 - Liz) 使用者輸入的地址
 
-  // filingFrequency: FILING_FREQUENCY; // Info: (20250423 - Liz) 申報頻率
-  // filingMethod: FILING_METHOD; // Info: (20250423 - Liz) 總繳種類
-  // declarantFilingMethod: DECLARANT_FILING_METHOD; // Info: (20250423 - Liz) 申報方式
+  // Info: (20250515 - Liz) 以下欄位為選填
+  filingFrequency?: FILING_FREQUENCY; // Info: (20250423 - Liz) 申報頻率
+  filingMethod?: FILING_METHOD; // Info: (20250423 - Liz) 總繳種類
+  declarantFilingMethod?: DECLARANT_FILING_METHOD; // Info: (20250423 - Liz) 申報方式
 
-  // declarantName: string; // Info: (20250423 - Liz) 申報人姓名
-  // declarantPersonalId: string; // Info: (20250423 - Liz) 申報人身分證字號
-  // declarantPhoneNumber: string; // Info: (20250423 - Liz) 申報人電話號碼
+  declarantName?: string; // Info: (20250423 - Liz) 申報人姓名
+  declarantPersonalId?: string; // Info: (20250423 - Liz) 申報人身分證字號
+  declarantPhoneNumber?: string; // Info: (20250423 - Liz) 申報人電話號碼
 
-  // agentFilingRole?: AGENT_FILING_ROLE; // Info: (20250423 - Liz) 申報代理人的角色，有三種：會計師(稅務代理人)、記帳士、記帳及報稅代理人
-  // licenseId?: string; // Info: (20250506 - Liz) 申報代理人的證書字號、登錄字號
+  agentFilingRole?: AGENT_FILING_ROLE; // Info: (20250423 - Liz) 申報代理人的角色，有三種：會計師(稅務代理人)、記帳士、記帳及報稅代理人
+  licenseId?: string; // Info: (20250506 - Liz) 申報代理人的證書字號、登錄字號
 }
 
 export interface IAccountBookWithTeam extends IAccountBook {
